Finished CSV file reading, ready to use games.java
